% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_authors.R
\name{add_authors}
\alias{add_authors}
\title{Adiciona o nome do autor ao título da espécie.}
\usage{
add_authors(species_name)
}
\arguments{
\item{species_name}{O nome do autor.}
}
\value{
O título da espécie com o nome do autor adicionado.
}
\description{
Adiciona o nome do autor ao título da espécie.
}
